+++ custom-rules/index.js
@@ -1,0 +1,11 @@
+/**
+ * Custom ESLint rules for @claxxon-lint
+ */
+
+import vueComponentNameThe from './vue-component-name-the.js';
+
+export default {
+  rules: {
+    'vue-component-name-the': vueComponentNameThe
+  }
+};